requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
efficientnet_pytorch
pandas
opencv-python
mtcnn
